requirements.txt
Flask==1.1.2
Flask_WTF==0.14.3
Flask_Bootstrap==3.3.7.1
requests==2.22.0
Pillow==8.1.0
